==> code/analysis/02b_GRCCA-genes/02_GRCCA-res-GSEA.qmd <==
---
title: "02b-02. Characterize the biological enrichments of GRCCA data using GSEA"
author: "Rachel Smith"
format: html
editor: visual
---

Run GSEA on GRCCA results. Note that the input data (cell types, GO, and risk gene benchmark lists) are loaded & collated in `setup.R`.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
base_dir <- "~/Documents/PhD/projects/sgacc_wgcna_grcca/"
scripts_dir <- paste0(base_dir, "code/analysis/")
```

## Setup and Load Data

```{r load_data}
source(paste0(scripts_dir, "load_WGCNA_res.R"))
load(paste0(objects_dir, "GRCCA_results.Rdata")) # df_lvs, df_y_res, df_x_res
```

## Get ranked gene list for GSEA

```{r rank_genes}
my_grcca_ranked <- df_x_res %>% 
    arrange(-pearsons_r) %>% 
    dplyr::select(ensembl_gene_id, pearsons_r) %>% 
    deframe
```

## GSEA Cell type

```{r cell_type}
df_grcca_fgsea_cell <- fgsea(pathways = cell_types, 
                          stats = my_grcca_ranked, 
                          eps = 0
)  %>% 
    as_tibble() %>% 
    dplyr::rename("cell_type" = "pathway") %>% 
    arrange(-abs(NES))
```

## GSEA GO

```{r go}
df_grcca_fgsea_go <- fgsea(pathways = go_pathways, 
                        stats = my_grcca_ranked, 
                        eps = 0
) %>% 
    as_tibble() %>% 
    dplyr::rename("go_id" = "pathway") %>% 
    arrange(-abs(NES)) %>% 
    left_join(df_go_terms) %>% # add names of GO terms
    dplyr::select(go_id, term, ontology, everything())
```

## GSEA Risk genes

```{r risk_genes}
df_grcca_fgsea_benchmark <- fgsea(pathways = benchmarking_lists, 
                               stats = my_grcca_ranked, 
                               eps = 0
) %>% 
    as_tibble() %>% 
    dplyr::rename("benchmark_list" = "pathway") %>% 
    arrange(-abs(NES))
```

## Identify WGCNA modules enriched for GRCCA genes

```{r FigS22.prep}
## Identify significant GRCCA genes (FDR < 0.05 & |z| > 2)
sig_grcca_genes <- df_x_res %>% 
    dplyr::filter(p_adj < 0.05 & abs(z_score) > 2) %>% 
    pull(ensembl_gene_id)
# n = 1211

## Hypergeometric test for significant overlap with modules
df_grcca_module_hypergeometric <- map_dfr(
    .x = names(module_colors),
    .f = ~ f_module_hypergeometric(
        gene_list = sig_grcca_genes, 
        gene_list_name = "", 
        wgcna_module = .x, 
        gene_universe = ensembl_gene_ids
    ) %>% 
        dplyr::select(-gene_list)
) %>% 
    mutate(p_adj = p.adjust(p_value, method = "fdr")) %>% 
    arrange(p_adj) %>% 
    mutate(module = factor(module, levels = names(module_colors))) %>% 
    
    # specify modules to label
    mutate(module_label = ifelse(p_adj < 0.05, paste0("N = ", overlap_n), ""))

## Load WGCNA GO enrichments results
prefix <- "08Mar2024_GENES_qSVAgeSexRaceGC" # Assuming prefix
soft_power <- 3 # Assuming soft_power
load(paste0(objects_dir, prefix, "_SIGNED_SFT", soft_power, "_GO_RES.RDS")) # df_mods_go
```

## Save

```{r save}
save(df_grcca_fgsea_cell, df_grcca_fgsea_go, df_grcca_fgsea_benchmark,
     file = paste0(objects_dir, "GRCCA_enrichment_benchmark_res.Rdata")
)
```

## Export GSEA results as table for supplement

```{r tableS3}
write_xlsx(
    list(
        "A.Risk genes" = df_grcca_fgsea_benchmark %>% 
            mutate(genes = map_chr(leadingEdge, ~ f_collapse_genes(.x, index = 30000))) %>% 
            dplyr::select(-leadingEdge) %>% unnest(cols = c(genes)),
        "B.Cell-type" = df_grcca_fgsea_cell %>% 
            mutate(genes = map_chr(leadingEdge, ~ f_collapse_genes(.x, index = 30000))) %>% 
            dplyr::select(-leadingEdge) %>% unnest(cols = c(genes)),
        "C.GO" = df_grcca_fgsea_go %>% 
            dplyr::filter(padj < 0.05) %>%
            mutate(genes = map_chr(leadingEdge, ~ f_collapse_genes(.x, index = 30000))) %>% 
            dplyr::select(-leadingEdge) %>% unnest(cols = c(genes))
    ), 
    path = paste0(tables_dir, "TableS3.GRCCA_GSEA_enrichments.xlsx")
)
```

## Plots

### Figure 3A \| GRCCA results are significantly enriched for SCZ risk genes

```{r Fig3a.plot.left, fig.height = 2, fig.width = 2.5}
## Prep
df_fig3a_left <- df_grcca_fgsea_benchmark %>% 
    mutate(class = ifelse(str_detect(benchmark_list, "SCZ"), "SCZ", "other") %>% factor(levels = c("SCZ", "other"))) %>% 
    mutate(benchmark_list = str_remove_all(benchmark_list, "SCZ|\\)|\\(") %>% 
               str_trim %>% 
               str_replace(", ", ",\n")
    ) %>% 
    mutate(label_color = ifelse(padj < 0.05, "white", "black")) %>% 
    mutate(box_color = ifelse(padj < 0.05, "black", "transparent")) %>% 
    mutate(label = round(NES, 2) %>% format(digits = 3),
           label = ifelse(padj < 0.05, paste0(label, "*"), label)
    )

## Plot
df_fig3a_left %>% 
    ggplot(aes(x = benchmark_list, y = 1)) +
    geom_tile(aes(fill = NES, color = I(box_color)), width = 0.95, height = 0.95, linewidth = 0.75) +
    geom_text(aes(label = label, color = I(label_color)), size = 3) +
    facet_wrap(vars(class), scales = "free", nrow = 2) +
    scale_fill_gradientn(colors = rev(brewer.rdbu(100)[1:50]), limits = c(1.0, 1.58)) +
    guides(fill = guide_colorbar(title.hjust = 0.5, title.position = "left")) +
    labs(x = NULL, y = NULL,
         title = "A | Risk gene enrichment") +
    theme(axis.text.y = element_blank(),
          axis.ticks = element_blank(),
          axis.line = element_blank(),
          axis.text.x = element_text(margin = margin(t = -3)),
          strip.background = element_rect(fill = "white", color = "black"),
          legend.margin = margin(r = -15),
          legend.position = "left",
          legend.justification = "left",
          legend.title = element_text(angle = 90),
          legend.key.height = unit(0.75, "cm"),
          legend.key.width = unit(0.15, "cm")
    )


## Save
map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(figures_dir, "3Aleft.Risk_gene_enrichment", .x), 
                width = 2.5, height = 2)
)
```

```{r Fig3a.plot.right, fig.width = 1.3, fig.height = 1.3}
## Prep

# Identify genes to label in right panel
top_genes <- df_x_res %>% 
    dplyr::filter(ensembl_gene_id %in% benchmarking_lists[["SCZ (common, broad)"]]) %>% 
    top_n(n = 6, wt = abs(pearsons_r)) %>% 
    pull(ensembl_gene_id)

# Subset X res for genes in SCZ common broad list for right panel
df_fig3a_right <- df_x_res %>% 
    dplyr::filter(ensembl_gene_id %in% benchmarking_lists[["SCZ (common, broad)"]]) %>% 
    mutate(significant = ifelse(p_adj < 0.05, 1, 0) %>% factor) %>% 
    arrange(significant, abs(pearsons_r)) %>% 
    mutate(label = ifelse(ensembl_gene_id %in% top_genes, gene_symbol, ""))


## Plot
scale_max <- df_x_res %>% pull(pearsons_r) %>% abs %>% max
jitter_pos <- position_jitter(width = 0.1, seed = 123)
df_x_res %>%
    ggplot(aes(x = pearsons_r)) +
    geom_vline(xintercept = 0, color = "darkgrey") +
    geom_density() +
    geom_point(data = df_fig3a_right, 
               mapping = aes(x = pearsons_r, y = -0.5, fill = pearsons_r, color = significant),
               pos = jitter_pos, shape = 21, size = 1
    ) +
    geom_text_repel(data = df_fig3a_right,
                    mapping = aes(x = pearsons_r, y = -0.5, label = label),
                    pos = jitter_pos, 
                    size = 2.5, box.padding = 0.75, max.overlaps = 100, 
                    force = 10, min.segment.length = 0
    ) +
    
    scale_fill_gradientn(colors = gene_weight_color_scale, 
                         limits = c(-scale_max, scale_max), guide = "none") +
    scale_color_manual(values = c("transparent", "black"), guide = "none") +
    labs(x = expression("Structure correlation (" * italic(r)[x] * ")"), y = NULL, 
         title = NULL) +
    theme(axis.text.y = element_blank(),
          axis.ticks.y = element_blank(),
          axis.text.x = element_text(size = 7),
          axis.title.x = element_text(size = 8),
          panel.border = element_rect(fill = "transparent", color = "black", size = 1),
          axis.line = element_blank()
    )

## Save
map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(figures_dir, "3Aright.Risk_gene_enrichment", .x), 
                width = 1.3, height = 1.3)
)
```

### Figure 3B \| GRCCA cell-type enrichments

```{r Fig3b.plot, fig.height = 2.25, fig.width = 2.5}
df_grcca_fgsea_cell %>% 
    mutate(sig = case_when(padj < 0.001 ~ "***", 
                           padj < 0.01 ~ "**", 
                           padj < 0.05 ~ "*", 
                           TRUE ~ "")) %>% 
    
    ggplot(aes(x = NES, y = reorder(cell_type, NES))) +
    geom_col(aes(fill = NES), color = "gray", shape = 21) +
    geom_text(aes(label = sig, x = NES), vjust = 0.75, size = 4, check_overlap = TRUE) +
    scale_fill_gradientn(colors = rev(brewer.rdbu(100)), limits = c(-3.39, 3.39)) +
    coord_cartesian(clip = "off") +
    labs(x = "NES", y = NULL,
         title = "B | Cell-type enrichment") +
    theme(legend.position = "none")

map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(figures_dir, "3B.cell_type_enrichment", .x), 
                width = 2.5, height = 2.25)
)
```

### Figure 3C \| GRCCA GO enrichments

```{r Fig3c.prep}
## Determine labels
fig3c_labels <- c(
    "cilium movement",
    #"motile cilium",
    "axoneme",
    #"axonemal microtubule",
    #"axoneme assembly",
    "adaptive immune response",
    #"outer dynein arm assembly",
    "MHC class II protein complex",
    #"T cell costimulation",
    #"MHC class II protein complex binding",
    "minus-end-directed microtubule motor activity",
    "toll-like receptor signaling pathway",
    #"dynein light intermediate chain binding",

    "syntaxin-1 binding",
    #"ubiquitin-dependent protein catabolic process",
    #"endosome to lysosome transport",
    #"cyleftlasmic translation",
    "regulation of autophagy",
    "postsynaptic density",
    "presynaptic endocytic zone membrane",
    "ATP binding",
    #"ATP hydrolysis activity",
    "protein ubiquitination"
)

## Filter for terms of interest & assign labels
df_fig3c <- df_grcca_fgsea_go %>% 
    dplyr::filter(padj < 0.05) %>% 
    mutate(direction = ifelse(NES < 0, "down", "up")) %>% 
    group_by(direction) %>% 
    mutate(label = ifelse(term %in% fig3c_labels, str_wrap(term, 25), ""))
```

```{r Fig3c.plot, fig.width = 6.5, fig.height = 4}
df_fig3c %>%
    ggplot(aes(x = padj, y = NES)) +
    geom_point(aes(size = size, fill = NES), color = "gray", shape = 21) +
    geom_text_repel(aes(label = label), size = 2.75, force = 20,
                    box.padding = 0.25, min.segment.length = 0, max.overlaps = 50) +
    geom_hline(yintercept = 0, color = "gray") +
    facet_wrap(vars(ontology)) +
    scale_fill_gradientn(colors = rev(brewer.rdbu(100)), limits = c(-3, 3), guide = "none") +
    scale_size_continuous(range = c(2, 6)) +
    scale_x_reverse(breaks = c(0, 0.03, 0.05)) +
    coord_cartesian(clip = "off") +
    guides(size = guide_legend(title = "n genes in path", 
                               title.hjust = 0.5, 
                               title.position = "top")) +
    labs(x = "Enrichment FDR", y = "NES",
         title = "C | Functional pathway enrichment") +
    theme(
          legend.justification = "center"
          legend.position = "bottom"
    )

## Save
map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(figures_dir, "3C.GO_enrichment", .x), 
                width = 6.5, height = 4)
)
```

### Figure S22A \| WGCNA modules enriched for GRCCA genes

```{r FigS22a.plot, fig.height = 5, fig.width = 2.15}
f_plot_module_hypergeometric(df_grcca_module_hypergeometric %>% 
                                 mutate(p_adj = ifelse(p_adj == 0, 10^-5, p_adj)) %>% 
                                 dplyr::filter(module != "geneM0"),
                             x_axis = "p-adj",
                             plot_title = "A | GRCCA significant modules",
                             #subset_x_labs = TRUE,
                             include_module_labels = TRUE,
                             module_text_size = 3,
                             legend_position = "none",
                             p_value_threshold = 0.05,
                             include_threshold_label = TRUE
)

## Save
map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(supp_figures_dir, "S22A.GRCCA_module_hypergeometric", .x), w
                idth = 2.15, height = 5)
)
```

### Figure S22B \| Direction of significant genes in left modules

```{r FigS22b.prep}
## Identify significant modules
sig_modules <- df_grcca_module_hypergeometric %>% 
    dplyr::filter(module != "geneM0" & p_adj < 0.05) %>% 
    pull(module) %>% sort()

## Identify panel sizes (number of significant genes in each module)
panel_widths <- df_x_res %>% 
    dplyr::filter(module %in% sig_modules & ensembl_gene_id %in% sig_grcca_genes) %>% 
    dplyr::count(module) %>% 
    pull(n)
```

```{r FigS22b.plot, fig.height = 5, fig.width = 2.15}
df_x_res %>% 
    dplyr::filter(module %in% sig_modules & ensembl_gene_id %in% sig_grcca_genes) %>% 
    mutate(weight = ifelse(significant == 1, weight, 0)) %>% 
    
    ggplot(aes(x = reorder_within(ensembl_gene_id, within = module, by = pearsons_r), 
               y = pearsons_r)) +
    geom_col(aes(fill = module, alpha = abs(pearsons_r))) +
    facet_wrap(vars(module), ncol = 1, scales = "free_y") +
    force_panelsizes(rows = panel_widths) +
    scale_fill_manual(values = module_colors) +
    scale_alpha_continuous(range = c(0.3, 1)) +
    labs(title = "B | Significant module genes",
         x = "Gene", y = expression("Structure correlation (" * italic(r)[x] * ")")) +
    guides(fill = guide_legend(nrow = 2)) +
    coord_flip() +
    theme(axis.text.y = element_blank(),
          axis.ticks.y = element_blank(),
          strip.text = element_blank(),
          legend.position = "none",
          panel.spacing = unit(0.3, "lines")
    )

## Save
map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(supp_figures_dir, "S22B.sig_module_direction", .x), 
                width = 2.15, height = 5)
)

```

### Figure S22C \| GO BP enrichments of significant modules

```{r FigS22c.prep}
# Filter for GRCCA modules of interest & take left 5 pathways for each
df_figS22c <- df_mods_go %>% 
    dplyr::filter(module %in% sig_modules) %>% 
    group_by(module) %>% 
    left_n(5, wt = -log10(pvalue)) %>% 
    mutate(Description = str_wrap(Description, width = 23.5))
```

```{r FigS22c.bubbleplot, fig.height = 5, fig.width = 2.15}
df_figS22c %>% 
    ggplot(aes(x = p.adjust, y = reorder_within(Description, -p.adjust, module))) +
    geom_point(aes(size = RichFactor, fill = module), shape = 21) +
    facet_wrap(vars(module), scales = "free_y", ncol = 1) +
    scale_fill_manual(values = module_colors, guide = "none") +
    scale_y_reordered() +
    scale_x_reverse(breaks = c(0.00005, 0.00)) +
    scale_size_continuous(range = c(1, 4)) +
    coord_cartesian(clip = "off") +
    labs(x = "Enrichment FDR", y = NULL,
         title = "C | Module enrichments") +
    theme(
        legend.position = c(-1.6, 1.05),
        legend.key.size = unit(0.2, "cm"),
        axis.text = element_text(size = 8.5)
    )

## Save
map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(suppl_figures_dir, "S22C.sig_GRCCA_module_enrichments", .x), 
                width = 2.15, height = 5)
)
```
