==> code/analysis/02b_GRCCA-genes/03_developmental-trajectories.qmd <==
---
title: "02b-03. Characterize normative developmental expression of GRCCA results deciles"
author: "Rachel Smith"
format: html
editor: visual
---

Characterize normative developmental trajectories of genes in each decile of GRCCA structure correlation results.

**Now in supplement only!**

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
base_dir <- "~/Documents/PhD/projects/sgacc_wgcna_grcca/"
```

## Setup

```{r load_data}
source(paste0(scripts_dir, "load_WGCNA_res.R"))
load(paste0(objects_dir, "GRCCA_results.Rdata")) # df_lvs, df_y_res, df_x_res
```

## Developmental trajectories (data)

```{r developmental_trajectories}
load(paste0(objects_dir, "objects/psychencode_development_expr_data.Rdata")) # df_psychencode_metadata, df_psychencode_expr


## Identify samples from region of interest
df_sample_window <- df_psychencode_metadata %>% 
    dplyr::filter(region_broad == "frontal_cortex")
frontal_cortex_samples <- df_sample_window%>% 
    pull(id)


## Filter expression data for samples and genes of interest
df_psychencode_expr_filt <- df_psychencode_expr %>% 
    dplyr::filter(GENE %in% ensembl_gene_ids) %>% 
    dplyr::select(GENE, all_of(frontal_cortex_samples)) %>% 
    dplyr::rename("ensembl_gene_id" = "GENE")


## Combine developmental expression data with GRCCA results (split into deciles)
df_psychencode_grcca <- df_psychencode_expr_filt %>% 
    pivot_longer(starts_with("HSB"), names_to = "id", values_to = "dev_expression") %>% 
    left_join(df_sample_window, by = join_by(id)) %>% 
    dplyr::select(window, id, regioncode, ensembl_gene_id, dev_expression) %>% 
    left_join(df_x_res %>% 
                  mutate(decile = ntile(-pearsons_r, 10) %>% factor)
    )

## Save for figures
save(df_psychencode_grcca,
     file = paste0(objects_dir, "GRCCA_developmental_expression.RDS")
)
```

## Plots

### Figure S23 \| Developmental expression trajectories of GRCCA genes

```{r FigS23.prep}
## Calculate the mean expression for each decile of GRCCA results across development
df_figS23 <- df_psychencode_grcca %>% #dplyr::filter(regioncode == "MFC") %>% 
    group_by(decile, window) %>% 
    summarise(median_expression = median(dev_expression),
              sd_expression = sd(dev_expression)
    )

## Rename windows with time period
psychencode_windows <- c("5-9",
                "12-13",
                "16-18",
                "19-22",
                "PCW35-PY0.3",
                "0.5-2.5",
                "2.8-10.7",
                "13-19",
                "21-64")
names(psychencode_windows) = 1:9  

## Define colors for deciles
decile_colors <- colorRampPalette(c(gene_weight_color_scale[5], "gray", gene_weight_color_scale[95]))(10) %>% rev
```

```{r FigS23, fig.height = 3, fig.width = 4}
df_figS23 %>% 
    ggplot(aes(x = window, y = median_expression, color = decile)) +
    geom_vline(xintercept = 5, color = "gray") +
    geom_smooth(method = "loess", se = FALSE) +
    scale_x_continuous(breaks = seq(1, length(psychencode_windows), 1),
                       labels = psychencode_windows) +
    scale_color_manual(values = decile_colors) +
    guides(color = guide_legend()) +
    labs(x = "← PCW | PY →", y = "Median expression",
         title = "Developmental expression trends") +
    theme(
        legend.key.size = unit(0.3, "cm"),
        axis.text.x = element_text(angle = 45, hjust = 1)
    )

## Save
map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(supp_figures_dir, "S23.developmental_trajectory", .x), 
                width = 4, height = 3)
)
```
