==> code/analysis/02b_GRCCA-genes/04_compare-to-DEGs.qmd <==
---
title: "02b-04. Assess the relationship between DEG test statistics and GRCCA structure correlations"
author: "Rachel Smith"
format: html
editor: visual
---

Characterize the relationship between DGE L2FC and GRCCA rx using permutation testing.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
base_dir <- "~/Documents/PhD/projects/sgacc_wgcna_grcca/"
```

## Setup and Load Data

```{r load_data}
source(paste0(scripts_dir, "load_WGCNA_res.R"))
load(paste0(objects_dir, "GRCCA_results.Rdata")) # df_lvs, df_y_res, df_x_res
load(paste0(objects_dir, "DE_results.Rdata")) # df_limma_res (and others)
```

## Combine GRCCA and DEG results

```{r combine_results}
df_de_grcca <- df_x_res %>% 
    left_join(df_limma_res, by = join_by(ensembl_gene_id, gene_symbol))

## Calculate the observed correlation between rx and DEG L2FC
cor.test(df_de_grcca$pearsons_r, df_de_grcca$log_fc)
```

## Run permutation testing

Resampling within module to determine the significance of the correlation

```{r permutation_test}
n_perm <- 10000
df_limma_grcca_cor_perm <- tibble()
set.seed(0406)
for (i in 1:n_perm) {
    
    if (i %% 100 == 0) {
        # print(paste0("Running permutation ", i))
    }
    
    # Permute LFC within module
    df_tmp <- df_de_grcca %>% 
        group_by(module) %>%
        mutate(permuted_r = sample(pearsons_r, replace = FALSE))
    
    # Calculate correlation between permuted data and Nirmala's LFC
    test <- cor.test(df_tmp$permuted_r, df_tmp$log_fc)
    correlation <- test$estimate
    p_value <- test$p.value
    
    # Append to tibble
    df_limma_grcca_cor_perm <- df_limma_grcca_cor_perm %>% 
        bind_rows(
            tibble(
                perm = i,
                correlation = correlation,
                p_value = p_value
            )
        )
    
}

## Calculate permutation p-values
de_limma_grcca_perm_p <- df_limma_grcca_cor_perm %>% 
    mutate(obs_corr = cor.test(df_de_grcca$pearsons_r, df_de_grcca$log_fc)$estimate) %>%
    summarise(
        perm_p = (sum(abs(correlation) >= abs(obs_corr)) + 1) / (n_perm + 1)
    )
# Pperm = 1*10^-4
```

## Plots

### Figure 4C \| GRCCA structure correlations (rx) vs DGE effect size

```{r Fig4c.prep}
## Combine GRCCA and DGE results, select risk (?) genes to label
genes_to_label <- c("BCL7A", "DNAH10", "BAI3", "PTK2B", "CSMD1", "FOXP1")
df_fig4c <- df_x_res %>% 
    left_join(df_limma_res, by = join_by(ensembl_gene_id, gene_symbol)) %>% 
    mutate(
        #risk_gene = ifelse(ensembl_gene_id %in% benchmarking_lists[["SCZ (common, broad)"]], "yes", "no"),
        risk_gene = ifelse(gene_symbol %in% genes_to_label, "yes", "no"),
        label = ifelse(gene_symbol %in% genes_to_label, gene_symbol, NA_character_)
    ) %>% 
    arrange(risk_gene)
```

```{r Fig4c.plot, fig.width = 2.17, fig.height = 2.0}
df_fig4c %>% 
    ggplot(aes(x = log_fc, y = pearsons_r)) +
    geom_hline(yintercept = 0, color = "gray") +
    geom_vline(xintercept = 0, color = "gray") +
    geom_point(aes(fill = pearsons_r + log_fc, color = risk_gene, size = risk_gene), 
               shape = 21, stroke = 0.25) +
    geom_text_repel(aes(label = label), min.segment.length = 0, size = 2.5) +
    geom_smooth(method = "lm", color = "black", linewidth = 0.25) +
    stat_cor(aes(label = after_stat(r.label)), label.sep = "\n", size = 3,
             label.y.npc = "top", vjust = -1) +
    scale_fill_gradientn(colors = gene_weight_color_scale, limits = c(-2.41, 2.41)) +
    scale_color_manual(values = c("yes" = "black", "no" = "transparent")) +
    scale_size_manual(values = c("yes" = 1, "no" = 0.5)) +
    coord_cartesian(clip = "off") +
    #scale_x_continuous(breaks = c(-0.2, 0, 0.2)) +
    labs(x = "log2(fold change)", 
         y = expression("Structure correlation (" * italic(r)[x] * ")"),
         title = "C | DGE vs GRCCA"
    ) +
    theme(legend.position = "none")

map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(figures_dir, "4C.gene_strucCor_vs_DE_stat", .x), 
                width = 2.17, height = 2.0)
)
```

### Figure 4F \| Compare DGE and GRCCA functional enrichments

```{r Fig4f.prep}
df_fig4f <- df_grcca_fgsea_go %>% 
    mutate(analysis = "GRCCA", .before = 1) %>% 
    bind_rows(df_de_fgsea_go %>% 
                  mutate(analysis = "DGE", .before = 1)
    ) %>% 
    mutate(class = ifelse(NES < 0, "NES < 0", "NES > 0")) %>% 
    filter(padj < 0.05) %>% 
    count(analysis, class) %>% 
    mutate(text_color = ifelse(analysis == "DGE", "white", "black"),
           analysis_color = ifelse(analysis == "GRCCA", "white", "black")
    )
```

```{r Fig4e.plot, fig.width = 2.17, fig.height = 1.50}
df_fig4f %>% 
    ggplot(aes(x = analysis, y = n)) +
    geom_rect(aes(xmin = -Inf, xmax = Inf, ymin = -Inf, ymax = Inf, fill = class), 
              alpha = 0.35) +
    geom_col(aes(fill = I(analysis_color)), color = "black") +
    geom_text(aes(label = n, color = I(text_color)), size = 3, vjust = 1.2) +
    facet_wrap(vars(class)) +
    scale_fill_manual(values = c("NES < 0" = gene_weight_color_scale[15], 
                                 "NES > 0" = gene_weight_color_scale[85])) +
    labs(x = NULL, y = "N pathways \n(FDR < 0.05)",
         title = "F | DGE vs GRCCA GO results"
    ) +
    theme(
        #panel.background = element_rect(fill = "transparent"),
        legend.position = "none",
        axis.text = element_text(size = 8)
    )
map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(figures_dir, "4F.GRCCA_vs_DGE_GO", .x), 
                width = 2.17, height = 1.50)
)
```
